Add files via upload
In this project, we aim to construct a real-time user behaviour analysis system for online shopping environment. We use an online retailer data to perform the experiments. In order to predict the purchasing intention of the visitor. Oversampling and feature selection pre-processing techniques are applied to improve the success rates and scalability of the algorithms.